django_blog/requirements.txt
# Django Core
Django==5.2.11
mysqlclient==2.2.8
asgiref==3.11.1
sqlparse==0.5.5
tzdata==2025.3

# Utilities (Important)
python-dotenv==1.2.1
requests==2.32.5
Pygments==2.19.2

# Optional (Keep only if you use them in code)
djangorestframework==3.16.1
django-filter==25.2

django-crispy-forms
crispy-bootstrap5

django-taggit

